First draft of the DSL. Not ready for production yet since JS generation is somewhat of an issue. Specified the API in a huge data structure for now.
If you know some better way to do this, let me know.

method_missing is used to check whether a method exists, validate
the arguments and add it to list. The list will be converted to valid
JS calls (based on use_async?) and rendered out for piwik_tracking_tag.

refs #7, #10